requirements.txt
chardet==3.0.4
click==7.1.1
cowsay==2.0.3
pyfiglet==0.8.post1
PyYAML==5.3.1